adjusted to optimize

+++ Priyas_Corner/post-optimization/compare1.py
@@ -104,7 +104,7 @@ def readFeatures (file):
     f = open(file)
     data = []
     for line in f:
-        data.append(int(line))
+        data.append(float(line))
 
     return data
